requirements.txt
flask>=3.0.0
google-genai>=0.8.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pypdf>=4.0.0
gunicorn>=21.2.0
flask-limiter>=3.5.0
